Update patch set 4
Patch Set 4:

Just tested, with the mp3 section commented out and the Google audio codecs import line at the bottom, audio still works. With the mp3 section not commented out and the line at the bottom, it doesn't work. Pushing a new PS in a couple minutes to reflect these changes.

Patch-set: 4